Restructured test directory structure, corrected namespace, also did some prettyfying
