+++ MonopolyUnitTests/CardTests/CardIntegrationTests.cs
@@ -12,8 +12,7 @@ using NUnit.Framework;
 using Ploeh.AutoFixture;
 using Ploeh.AutoFixture.AutoMoq;
 
-
-namespace MonopolyUnitTests
+namespace MonopolyUnitTests.CardTests
 {
     [TestFixture]
     class CardIntegrationTests
