+++ MonopolyUnitTests/CardTests/DeckUnitTests.cs
@@ -5,7 +5,7 @@ using NUnit.Framework;
 using Ploeh.AutoFixture;
 using Ploeh.AutoFixture.AutoMoq;
 
-namespace MonopolyUnitTests.TestClasses
+namespace MonopolyUnitTests.CardTests
 {
     [TestFixture]
     class DeckUnitTests
