+++ MonopolyUnitTests/HandlerTests/TurnHandlerIntegrationTests.cs
@@ -9,7 +9,7 @@ using NUnit.Framework;
 using Ploeh.AutoFixture;
 using Ploeh.AutoFixture.AutoMoq;
 
-namespace MonopolyUnitTests.TestClasses
+namespace MonopolyUnitTests.HandlerTests
 {
     [TestFixture]
     class TurnHandlerTests
